2003-02-11  Paolo Carlini  <pcarlini@unitus.it>
	PR libstdc++/9320
	* include/ext/stdio_filebuf.h
	(stdio_filebuf(int, std::ios_base::openmode, bool, int_type),
	stdio_filebuf(std::__c_file*, std::ios_base::openmode, int_type)):
	Change to take a __size parameter of type size_t, not
	of type (template parameter dependent) int_type.
	* src/ios.cc (ios_base::Init::_S_ios_create): Change type of
	size vars to size_t.
	* testsuite/ext/stdio_filebuf.cc: Add.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@62691 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libstdc++-v3/src/ios.cc
@@ -159,11 +159,12 @@ namespace std
   void
   ios_base::Init::_S_ios_create(bool __sync)
   {
-    int __out_size = __sync ? 0 : static_cast<int>(BUFSIZ);
+    size_t __out_size = __sync ? 0 : static_cast<size_t>(BUFSIZ);
 #ifdef _GLIBCPP_HAVE_ISATTY
-    int __in_size = (__sync || isatty (0)) ? 1 : static_cast<int>(BUFSIZ);
+    size_t __in_size =
+      (__sync || isatty (0)) ? 1 : static_cast<size_t>(BUFSIZ);
 #else
-    int __in_size = 1;
+    size_t __in_size = 1;
 #endif
 
     // NB: The file globals.cc creates the four standard files
